Move Xtx into a submodule
